[ship] MSVC build fixes + Windows TLS CA + live OpenAI stream running
- Fix circular include: split OutputSink.h (abstract only, fwd decl) from
  OutputSinkImpl.h (concrete CSV/JSON/Memory sinks with full TradeSignal)
- Rename emit() param signal→sig to avoid POSIX signal() macro collision
- Add ssize_t typedef via BaseTsd.h in LLMStreamClient.h for MSVC
- Switch compiler flags to /W3 /WX on MSVC (was -Wall -Wextra -Werror GCC-only)
- Fix LatencyController implicit int64→double cast (C4244 as error)
- Load Windows system CA store into OpenSSL SSL_CTX so cert verify passes
  against api.openai.com (vcpkg OpenSSL has no built-in CA bundle)
- MetricsLogger: use unique logger names + drop() to survive re-init; no throw
- main: parse --stream before config load; default to memory stream when no
  config file found (avoids FileNotFound crash on tokens.txt)
- Add .gitignore for build/ artifacts

Engine confirmed running live: gpt-4o streaming, 48 TPS, 300+ signals/min,
risk gate PASS/BLOCK firing, dedup active, pressure system online.

+++ src/LatencyController.cpp
@@ -70,7 +70,7 @@ LatencyController::LatencyStats LatencyController::get_stats() const {
             auto mean = stats.avg_latency.count();
             double variance = 0.0;
             for (const auto& sample : samples_copy) {
-                double diff = sample.count() - mean;
+                double diff = static_cast<double>(sample.count()) - static_cast<double>(mean);
                 variance += diff * diff;
             }
             stats.jitter_ms = std::sqrt(variance / samples_copy.size()) / 1000.0;
